Fixing issue #1, changing datatype from float to decimal.

+++ BitfinexApiSharp/BitfinexClientSharp/Dtos/TickerResponse.cs
@@ -4,15 +4,15 @@
     {
         public Pair Pair { get; set; }
         public string ChannelId { get; set; }
-        public float Bid { get; set; }
-        public float BidSize { get; set; }
-        public float Ask { get; set; }
-        public float AskSize { get; set; }
-        public float DailyChange { get; set; }
-        public float DailyChangePercentage { get; set; }
-        public float LastPrice { get; set; }
-        public float Volume { get; set; }
-        public float High { get; set; }
-        public float Low { get; set; }
+        public decimal Bid { get; set; }
+        public decimal BidSize { get; set; }
+        public decimal Ask { get; set; }
+        public decimal AskSize { get; set; }
+        public decimal DailyChange { get; set; }
+        public decimal DailyChangePercentage { get; set; }
+        public decimal LastPrice { get; set; }
+        public decimal Volume { get; set; }
+        public decimal High { get; set; }
+        public decimal Low { get; set; }
     }
 }
